\name{StatComp20047-package}
\alias{StatComp20047-package}
\alias{StatComp20047}
\docType{package}
\title{
  Several algorithms for ML and 20047's all homework for 'Statistical Computing' course. 
}
\description{
  This package provides several functions for machine learning and 20047' all homework for 'Statistical Computing' course.
}
\details{
  Gradient descent algorithm and stochastic gradient descent algorithm are commonly used for classification and regression in machine learning.
}
\author{
Jianyu Liu

Maintainer: Jianyu Liu <liu06@mail.ustc.edu.cn>
}

\keyword{ package }

